+++ components/common/Util.ts
@@ -47,3 +47,15 @@ export const renderMdToHtml = (raw: string, escapeFrontMatter = false): string =
   }
   return mdParser.render(raw)
 }
+
+export const getCurrentDateTime = (): number => {
+  return Date.now()
+}
+
+export const showCurrentDateTime = (date: number): string => {
+  return new Date(date).toLocaleString()
+}
+
+export const getCurrentWeek = (): string => {
+  return 'week 1'
+}
