+++ pages/api/auth/[...nextauth].tsx
@@ -1,22 +1,7 @@
 import NextAuth from 'next-auth'
 import Providers from 'next-auth/providers'
 import { FirebaseAdapter } from '@next-auth/firebase-adapter'
-
-import firebase from 'firebase/app'
-import 'firebase/firestore'
-
-const firebaseConfig = {
-  apiKey: process.env.FIREBASE_API_KEY,
-  authDomain: process.env.FIREBASE_AUTH_DOMAIN,
-  databaseURL: process.env.FIREBASE_DATABASE_URL,
-  projectId: process.env.FIREBASE_PROJECT_ID,
-  storageBucket: process.env.FIREBASE_STORAGE_BUCKET,
-  messagingSenderId: process.env.FIREBASE_MESSAGING_SENDER_ID,
-  appId: process.env.FIREBASE_APP_ID,
-  measurementId: process.env.FIREBASE_MEASUREMENT_ID,
-}
-
-const firestore = (firebase.apps[0] ?? firebase.initializeApp(firebaseConfig)).firestore()
+import { firestore } from '../../../components/utils/firebase'
 
 export default NextAuth({
   // Configure one or more authentication providers
